misc: mods: fix potential integer overflow
Bug 3461002
CID 10138562

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2729925
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/misc/mods/mods_mem.c
@@ -2258,7 +2258,7 @@ static void clear_contiguous_cache(struct mods_client *client,
 	}
 #endif
 
-	d_size = 4 << d_line_shift;
+	d_size = (u64)4 << d_line_shift;
 	cur = virt_start & ~(d_size - 1);
 	do {
 		asm volatile("dc civac, %0" : : "r" (cur) : "memory");
